requirements.txt
django
mysqlclient
djangorestframework
drf-nested-routers
django-cors-headers
gunicorn
whitenoise
python-dotenv
pytest
model_bakery
pytest-django